add a cheesy test suite
It could be a lot fancier, but it should be enough for now.  Fixing
utf8, tab expansion, and escape sequences in the input will be nearly
impossible without some sort of regression tests.

tc5 uses a fancy Thai character that doesn't seem to render quite right
under Vim in my OS X Terminal.app.  I'm not sure if I've botched the
character somehow or if OS X is just buggy about this...