FIX: Don't invoke tput if --columns is passed
Under certain circumstances (e.g. when run as a cron job), tput is not able to determine the columns (tput: unknown terminal "unknown"). For that, the columns can be explicitly set via --columns=N, and nowrap can still do its job. However, the tput lookup is still performed, and results in an error message (which even triggers an email for the cron job). Lookup of available terminal columns should only happen if the columns are not specified on the command-line; this also slightly improves performance.
Move the $column initialization to after the command-line argument parsing.