FIX: Runtime error if input contains a standalone escape
An escape character that is not part of an ANSI escape sequence (i.e. does not match $ESCAPE_SEQUENCE_PATTERN) produces undef as the submatch because the success of the pattern matching isn't checked.
> $ nowrap
> foo^[bar
> Use of uninitialized value $esc_seq in subtraction (-) at nowrap line 147, <> line 1.
> Use of uninitialized value in addition (+) at nowrap line 148, <> line 1.
> Use of uninitialized value $append in concatenation (.) or string at nowrap line 171, <> line 1.
> Use of uninitialized value $esc_seq in subtraction (-) at nowrap line 147, <> line 1.
> Use of uninitialized value in addition (+) at nowrap line 148, <> line 1.
> Use of uninitialized value $append in concatenation (.) or string at nowrap line 171, <> line 1.

Move the pattern matching into the condition, so that the branch only is taken if the pattern actually matches.